requirements.txt
python-telegram-bot~=20.9
requests
python-dotenv